backend/requirements.txt
fastapi
uvicorn[standard]
openai
qdrant-client
psycopg2-binary
alembic
beautifulsoup4
sqlalchemy
pydantic
python-dotenv